requirements.txt
pdfplumber==0.10.3
PyMuPDF==1.23.8
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
spacy==3.7.2
transformers==4.36.2
torch>=2.2.0
scikit-learn==1.3.2
python-dotenv==1.0.0
pymongo==4.6.1
psycopg2-binary==2.9.9
regex==2023.12.25